Removing workflows, pagerduty and statuspage as they'll become recipes/plugins in future.
