+++ response/apps.py
@@ -6,13 +6,10 @@ class ResponseConfig(AppConfig):
 
     def ready(self):
         from .slack import (settings,
-                       signals,
-                       action_handlers,
-                       event_handlers,
-                       incident_commands,
-                       keyword_handlers,
-                       incident_notifications,
-                       dialog_handlers,
-                       workflows)
-        if settings.PAGERDUTY_ENABLED:
-            from .slack.workflows import pagerduty
+                            signals,
+                            action_handlers,
+                            event_handlers,
+                            incident_commands,
+                            keyword_handlers,
+                            incident_notifications,
+                            dialog_handlers)
