--- response/pagerduty/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from response.pagerduty.models import Escalation
-
-# Register your models here.
-admin.site.register(Escalation)
