+++ response/slack/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from response.slack.models import HeadlinePost, CommsChannel, Notification, UserStats, PinnedMessage, Workflow, WorkflowAdmin
+from response.slack.models import HeadlinePost, CommsChannel, Notification, UserStats, PinnedMessage
 
 
 # Register your models here.
@@ -8,4 +8,3 @@ admin.site.register(CommsChannel)
 admin.site.register(Notification)
 admin.site.register(UserStats)
 admin.site.register(PinnedMessage)
-admin.site.register(Workflow, WorkflowAdmin)
