--- response/slack/workflows/__init__.py
@@ -1,1 +1,0 @@
-from .statuspage import *
