--- response/slack/workflows/statuspage/__init__.py
@@ -1,1 +1,0 @@
-from .workflow import *
