--- response/slack/workflows/statuspage/constants.py
@@ -1,3 +1,0 @@
-OPEN_STATUS_PAGE_DIALOG = 'open-status-page-dialog'
-STATUS_PAGE_UPDATE = 'status-page-update'
-STATUS_PAGE_COMMANDS = ['statuspage', 'sp']
